requirements.txt
Django==1.9
whitenoise==3.2
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.1
behave-django
selenium